requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Testing==0.6.2
funcsigs==1.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mock==2.0.0
pbr==3.1.1
python-dotenv==0.6.5
requests==2.18.4
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2
